Change default var values
Reasons:

fullbrightmodels: It can be very hard to see dark players. People should
not be able to hide by selecting a dark color.

firstperson*depth: It was hard to understand the position and size of
your own body when it never seemed to touch anything. This change makes
it visible when you touch something.