Fix loadweap question bug
The bug was that if you change your loadout and then go back to the
profile menu and cancel without changing anything, you will get a
question about whether you want to save changes to loadweap.

This commit also removes variable loadweapall that was not needed.
The part that fixed the bug was removing the (< $q 0) check.