Smoother fades on top of fades
If you opened console or compass while having a menu open, the
background flashed.  What caused this problem was the combination of
the if statement `a < colour[i]` and the multiplication-assign after
it.  The are two ways to solve this, with different results; remove
one of them.

If I remove the if statement, the fades always multiply, which causes
the background to get very dark when there are multiple fades.

If I remove the multiplication, which was done in this commit, the
fades don't multiply any more so the background doesn't get darker
than it would if there was only one fade.  The only problem is that
now the second fade is a litte bit delayed, because it doesn't start
to change anything until it is darker than the current fade, because
of the remaining if statement.

This commit also fixes the style of one line to be the same as the
other similar ones.  This is not related to the fade issue.