Always reset cursor when opening compass menu
If you opened a compass menu while having another menu open, the
cursor would not start in the center.

The point of a compass menu is to quickly be able to select an option
based on the position of it.  This requires the mouse and the options
to always be in the same position relative to each other.